requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
django-model-utils==4.0.0
django-tinymce4-lite==1.8.0
djangorestframework==3.12.2
docopt==0.6.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
Js2Py==0.70
jsmin==2.2.2
more-itertools==8.7.0
packaging==20.4
Pillow==8.0.1
pipenv==2020.8.13
plotly==4.12.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
pySmartDL==1.3.4
pytz==2020.1
requests==2.25.1
retrying==1.3.3
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.1
tzlocal==2.1
urllib3==1.26.3
virtualenv==20.1.0
virtualenv-clone==0.5.4
whitenoise==5.2.0
xmltodict==0.12.0
